Second commit

+++ src/main/java/com/felinos/microservice/repository/OrderRepository.java
@@ -5,8 +5,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface OrderRepository extends JpaRepository <Order, Long> {
+public interface OrderRepository extends JpaRepository <Order, Integer> {
 
-      public Order findById(long id);
+      public Order findById(int id);
     }
